admin - Ordercontroller_getById

+++ jcart-administration/src/main/java/com/gy/jcartadministration/controller/OrderController.java
@@ -32,7 +32,7 @@ public class OrderController {
     }
 
     @GetMapping("/getById")
-    public OrderShowOutDTO getById(@RequestParam Integer orderId) {
+    public OrderShowOutDTO getById(@RequestParam Long orderId) {
         OrderShowOutDTO orderShowOutDTO = orderService.getById(orderId);
         return orderShowOutDTO;
     }
